+++ src/components/ConnectionInfo.js
@@ -162,8 +162,11 @@ class ConnectionInfo extends Component {
           window.location.reload();
         })
         .catch(res=>{
-          localStorage.clear();
-            window.location.href = "/login";
+          if(res.status === 401)
+                {
+                  localStorage.clear();
+                window.location.href = "/login";
+                }
         });
     }
   };
